Make sure the initial network is a giant component
In our theory, the initial network should be a giant component, so we should tell wether it is before we start attacking it. I name it with a suffix 'First', because I've a better way to improve it in later edition, but before attacking , this is the only way we can guarantee the precondition.